Add files via upload
I have used my Taylor expansion code in quantum optics for the calculation of an image called the ghost image. This code demonstrates immense mathematical detail, as well as visualizations of 3D functions, and physical processes. If a detailed understanding of my code or research is required, please contact me and/or request my master's thesis.